Split importing into classes
