+++ src/Analysis/Engine/Impl/AnalysisUnit.cs
@@ -109,8 +109,8 @@ namespace Microsoft.PythonTools.Analysis {
         }
 
         internal void Enqueue() {
-            if (!ForEval && !IsInQueue && !_suppressEnqueue) {
-                State.Queue.Append(this);
+            if (!ForEval && !IsInQueue && !_suppressEnqueue && NeedsAnalysis()) {
+                State.Queue.Enqueue(this);
                 AnalysisLog.Enqueue(State.Queue, this);
                 IsInQueue = true;
 
@@ -342,6 +342,9 @@ namespace Microsoft.PythonTools.Analysis {
         internal virtual ILocationResolver AlternateResolver => null;
 
         ILocationResolver ILocationResolver.GetAlternateResolver() => AlternateResolver;
+
+        private bool NeedsAnalysis() 
+            => _analysisCount == 0 || _scope != null && _scope.AllVariables.Select(k => k.Value).Any(v => !v.HasTypes);
     }
 
     class ClassAnalysisUnit : AnalysisUnit {
